move branchstore file
